+++ src/compas/scene/geometryobject.py
@@ -18,15 +18,28 @@ class GeometryObject(SceneObject):
     ----------
     geometry : :class:`compas.geometry.Geometry`
         The geometry object associated with the scene object.
-    color : :class:`compas.colors.Color`
-        The color of the object.
+    pointcolor : :class:`compas.colors.Color`
+        The color of the points.
+    linecolor : :class:`compas.colors.Color`
+        The color of the lines or curves.
+    surfacecolor : :class:`compas.colors.Color`
+        The color of the surfaces.
+    pointsize : float
+        The size of the points.
+    linewidth : float
+        The width of the lines or curves.
 
     """
 
-    color = ColorAttribute(default=None)
+    pointcolor = ColorAttribute()
+    linecolor = ColorAttribute()
+    surfacecolor = ColorAttribute()
 
     def __init__(self, geometry, **kwargs):
         super(GeometryObject, self).__init__(item=geometry, **kwargs)
         self.geometry = geometry
-        if "color" in kwargs:
-            self.color = kwargs["color"]
+        self.pointcolor = kwargs.get("pointcolor", self.color)
+        self.linecolor = kwargs.get("linecolor", self.color)
+        self.surfacecolor = kwargs.get("surfacecolor", self.color)
+        self.pointsize = kwargs.get("pointsize", 1.0)
+        self.linewidth = kwargs.get("linewidth", 1.0)