+++ src/compas/scene/sceneobject.py
@@ -4,6 +4,7 @@ from __future__ import print_function
 
 from abc import abstractmethod
 from .descriptors.protocol import DescriptorProtocol
+from .descriptors.color import ColorAttribute
 from .context import clear
 from .context import get_sceneobject_cls
 
@@ -21,11 +22,19 @@ class SceneObject(object):
     ----------
     guids : list[object]
         The GUIDs of the items drawn in the visualization context.
+    transformation : :class:`compas.geometry.Transformation`
+        The transformation matrix of the scene object.
+    color : :class:`compas.colors.Color`
+        The color of the object.
+    opacity : float
+        The opacity of the object.
 
     """
 
     # add this to support the descriptor protocol vor Python versions below 3.6
     __metaclass__ = DescriptorProtocol
+
+    color = ColorAttribute()
 
     def __new__(cls, item, **kwargs):
         sceneobject_cls = get_sceneobject_cls(item, **kwargs)
@@ -35,6 +44,8 @@ class SceneObject(object):
         self._item = item
         self._transformation = None
         self._guids = None
+        self.color = kwargs.get("color", self.color)
+        self.opacity = kwargs.get("opacity", 1.0)
 
     @property
     def guids(self):
@@ -42,14 +53,6 @@ class SceneObject(object):
 
     @property
     def transformation(self):
-        """The transformation matrix of the scene object.
-
-        Returns
-        -------
-        :class:`Transformation` or None
-            The transformation matrix.
-
-        """
         return self._transformation
 
     @transformation.setter